model_docker/pyproject.toml
[project]
name = "biooton"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "air>=0.35.0",
    "cython>=3.2.4",
    "dbfread>=2.0.7",
    "fastparquet==2024.5.0",
    "ffmpeg-python>=0.2.0",
    "geopandas>=1.1.1",
    "geostatspy>=0.0.79",
    "hyperopt>=0.2.7",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "multiprocess>=0.70.18",
    "noisereduce>=3.0.3",
    "numpy==2.3.5",
    "optuna>=4.6.0",
    "pandas>=2.3.3",
    "paramiko>=4.0.0",
    "pathos>=0.3.4",
    "plotly>=6.5.2",
    "pycirclize>=1.10.1",
    "rasterio>=1.4.4",
    "ray[tune]>=2.53.0",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.6",
    "tensorboard>=2.20.0",
    "torchaudio>=2.9.1",
    "torchcodec>=0.9.1",
    "torchmetrics>=1.8.2",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
]
